created Plane object (3rd level) and tested pointer casting. We will have a vector of GenericObjects, but we need to be able to access GeometricObject-type object attributes like ambient, diffuse, specular, shinyness

+++ asg3/VS2013_Solution/Assignment3/GeometricObject.h
@@ -18,7 +18,7 @@ public:
 	virtual glm::vec3 getAmb() const { return amb; };
 	virtual glm::vec3 getDif() const { return dif; };
 	virtual glm::vec3 getSpe() const { return spe; };
-	virtual float getShi() const { return shi; };
+	virtual double getShi() const { return shi; };
 
 	virtual void setAmb(glm::vec3 newAmb) { amb = newAmb; }
 	virtual void setDif(glm::vec3 newDif) { dif = newDif; }
